[RHTAP-1069] Release controller refactoring
